refactor: Get only one file with new modifications
